+++ src/main.jsx
@@ -1,6 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { createRoot } from "react-dom/client";
-import { App } from './components/App/app';
+import { App } from "./App";
 
 const root = document.getElementById("root");
 const reactRoot = createRoot(root);
